+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - introduced a new command `bit untag` for reverting un-exported tags. 
+- init local scope inside .git
 
 ## [0.12.5] - 2018-02-06